Fail earlier if kerberos auth fails
Before this patch, if the auth request failed, we would carry on
anyway and not find out until we tried to do something else. Now it
should raise an error if there's a probelm with the auth request.

Also, a slightly related change to move a commented out print
statement, useful for debugging an update failures, so it will still
run if there is a failure.